Refactor old ``Hook`` to ``Property``
